+++ src/components/TodoList.tsx
@@ -1,0 +1,56 @@
+import { useContext } from 'react';
+import { DispatchContext, StateContext } from '../context/Store';
+
+export const TodoList = () => {
+  const state = useContext(StateContext);
+  const dispatch = useContext(DispatchContext);
+
+  const filteredTodos = state.todos.filter(todo => {
+    switch (state.filter) {
+      case 'active':
+        return !todo.completed;
+      case 'completed':
+        return todo.completed;
+      case 'all':
+      default:
+        return true;
+    }
+  });
+
+  return (
+    <section className="todoapp__main" data-cy="TodoList">
+      {filteredTodos.map(todo => (
+        <div
+          key={todo.id}
+          data-cy="Todo"
+          className={`todo ${todo.completed ? 'completed' : ''}`}
+        >
+          <label className="todo__status-label">
+            <input
+              data-cy="TodoStatus"
+              type="checkbox"
+              className="todo__status"
+              checked={todo.completed}
+              onChange={() =>
+                dispatch({ type: 'markCompleted', payload: todo.id })
+              }
+            />
+          </label>
+
+          <span data-cy="TodoTitle" className="todo__title">
+            {todo.title}
+          </span>
+
+          <button
+            type="button"
+            className="todo__remove"
+            data-cy="TodoDelete"
+            onClick={() => dispatch({ type: 'delete', payload: todo.id })}
+          >
+            ×
+          </button>
+        </div>
+      ))}
+    </section>
+  );
+};
